refactor: extract UserAccountsTab from admin index
Move the user accounts tab render function into a dedicated
component that receives all needed data and callbacks as props.
